requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
Django==4.2
dotenv==0.9.9
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
paho-mqtt==2.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.0.0
python-dotenv==1.1.0
redis==5.2.1
service-identity==24.2.0
setuptools==79.0.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
zope.interface==7.2
black==24.10.0
